Network Exception handle
